[#1]JHESMain add tab control
JHESMain add tab control
tab1 = singleXMLsend tab
tab2 = scenario tab
