+++ JHEquipSim/JHEquipSim/Views/JHESMain.resx
